Port back to linux again... Ugh, you wouldn't think you would have to touch so many files.
jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@582 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/api.c
@@ -22,6 +22,7 @@
 #include <fcntl.h>
 #include <sys/types.h>
 #include <sys/stat.h>
+#include <signal.h>
 #include <netinet/in.h>
 #include <string.h>
 #include <assert.h>
@@ -1252,7 +1253,7 @@ OWPReadDataHeader(
 		return 0;
 	}
 
-	if(stat_buf.st_size < sizeof(magic)+sizeof(ver)+sizeof(hlen)){
+	if(stat_buf.st_size < (off_t)(sizeof(magic)+sizeof(ver)+sizeof(hlen))){
 		OWPError(ctx,OWPErrFATAL,OWPErrUNKNOWN,
 				"OWPReadDataHeader:Invalid owp file");
 		return 0;
